chore: re-organize constants and types
