+++ src/components/member/member-profile-form.tsx
@@ -5,6 +5,8 @@ import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Button } from "@/components/ui/button";
 import { Textarea } from "@/components/ui/textarea";
+import { studyPrograms } from "@/lib/constants/study-programs";
+import type {ProfileFormData} from "@/types/profile"
 import {
   Select,
   SelectContent,
@@ -15,39 +17,6 @@ import {
 import { useToast } from "@/hooks/use-toast";
 import { useMemberProfile, useUpdateMemberProfile } from "@/hooks/member";
 
-interface ProfileFormData {
-  firstName: string;
-  lastName: string;
-  email: string;
-  university: string;
-  programme: string;
-  graduationYear: string;
-  githubLink: string;
-  linkedinLink: string;
-  aboutMe: string;
-}
-
-const studyPrograms = [
-  "Machine Learning",
-  "Applied Mathematics",
-  "Bio Technology",
-  "Engineering Physics",
-  "Computer Science",
-  "Electrical Engineering",
-  "Industrial Management",
-  "Information and Communication Technology",
-  "Chemical Science and Engineering",
-  "Mechanical Engineering",
-  "Mathematics",
-  "Material Science and Engineering",
-  "Medical Engineering",
-  "Environmental Engineering",
-  "The Built Environment",
-  "Technology and Economics",
-  "Technology and Health",
-  "Technology and Learning",
-  "Technology and Management",
-];
 
 export function MemberProfileForm() {
   const { toast } = useToast();
