+++ src/hooks/admin.ts
@@ -1,17 +1,10 @@
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
+import type {AdminProfileData,UpdateAdminUserProfileData, AdminUser} from "@/types/admin"
 
 const API_URL = `${
   process.env.NEXT_PUBLIC_API_URL || "http://localhost:8080"
 }/api/v1`;
 
-export interface AdminUser {
-  user_id: string;
-  email: string;
-  provider: string;
-  created_at: string;
-  updated_at: string;
-  roles: string[];
-}
 
 async function fetchAllUsers(): Promise<AdminUser[]> {
   const response = await fetch(`${API_URL}/admin/users`, {
@@ -68,29 +61,7 @@ export function useDemoteAdmin() {
   });
 }
 
-export interface AdminProfileData {
-  first_name: string;
-  last_name: string;
-  email: string;
-  university?: string;
-  programme?: string;
-  graduation_year?: number;
-  github_link?: string;
-  linkedin_link?: string;
-  about_me?: string;
-}
 
-export interface UpdateAdminUserProfileData {
-  firstName: string;
-  lastName: string;
-  email: string;
-  university?: string;
-  programme?: string;
-  graduationYear?: number;
-  githubLink?: string;
-  linkedinLink?: string; // PUT expects lowercase 'i'
-  aboutMe?: string;
-}
 
 // Fetch a single profile as an admin
 async function fetchAdminUserProfile(
