+++ src/lib/constants/study-programs.ts
@@ -1,0 +1,21 @@
+export const studyPrograms = [
+  "Machine Learning",
+  "Applied Mathematics",
+  "Bio Technology",
+  "Engineering Physics",
+  "Computer Science",
+  "Electrical Engineering",
+  "Industrial Management",
+  "Information and Communication Technology",
+  "Chemical Science and Engineering",
+  "Mechanical Engineering",
+  "Mathematics",
+  "Material Science and Engineering",
+  "Medical Engineering",
+  "Environmental Engineering",
+  "The Built Environment",
+  "Technology and Economics",
+  "Technology and Health",
+  "Technology and Learning",
+  "Technology and Management",
+];
